Fix a potential crash when conn of a delay is closed
If the connection is closed by the client while we are
waiting for the delay to finish, then the C cleanup
invalidates the request object, and the R processsing
loop must finish without performing for operations on
the request.

I also added a safety check, in case another operation
is attempted on the closed connection. Now the cleanup
also cleans up the external ptr in the request, to be
safe, and every operation checks that this external ptr
(referring to the mg connection) is valid.